requirements.txt
Django==4.2.13
Pillow==10.3.0
rembg==2.0.57
onnxruntime==1.18.0
